Fix a bug where VS2015 cannot compile

+++ src/Sim/Predictor/BPred/PHT.cpp
@@ -54,7 +54,7 @@ void PHT::Initialize(InitPhase phase)
         LoadParam();
         u8 max = (1 << m_counterBits) - 1;
         m_table.construct(
-            1 << m_entryBits,   // size
+            (u64)1 << m_entryBits,   // size
             (max + 1) / 2,      // init
             0,                  // min
             max,                // max
